# venture-forge/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Venture Forge Product API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ventures/{venture_id}/pipelines": {
      "post": {
        "summary": "Create Pipeline",
        "operationId": "create_pipeline_api_v1_ventures__venture_id__pipelines_post",
        "parameters": [
          {
            "name": "venture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venture Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ventures/{venture_id}/agent-runs": {
      "post": {
        "summary": "Enqueue",
        "operationId": "enqueue_api_v1_ventures__venture_id__agent_runs_post",
        "parameters": [
          {
            "name": "venture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venture Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ventures/{venture_id}/agent-runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_v1_ventures__venture_id__agent_runs__run_id__get",
        "parameters": [
          {
            "name": "venture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venture Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ventures/{venture_id}/agent-runs/{run_id}/review": {
      "post": {
        "summary": "Review",
        "operationId": "review_api_v1_ventures__venture_id__agent_runs__run_id__review_post",
        "parameters": [
          {
            "name": "venture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venture Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentReview"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ventures/{venture_id}/agent-runs/{run_id}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_api_v1_ventures__venture_id__agent_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "venture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venture Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AgentRequest": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "agent_id": {
            "type": "string",
            "enum": [
              "home",
              "research",
              "market",
              "customer",
              "competitor",
              "model",
              "finance",
              "experiment",
              "simulation",
              "academy",
              "ecosystem",
              "investor",
              "passport"
            ],
            "title": "Agent Id"
          },
          "objective": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Objective"
          },
          "hypothesis_id": {
            "type": "string",
            "title": "Hypothesis Id"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Evidence Ids"
          },
          "artifact_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 40,
            "title": "Artifact Ids"
          },
          "experiment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Experiment Ids"
          },
          "stage_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "AUTO",
              "RULE",
              "MODEL"
            ],
            "title": "Mode",
            "default": "AUTO"
          },
          "allow_model_processing": {
            "type": "boolean",
            "title": "Allow Model Processing",
            "default": false
          },
          "data_policy": {
            "type": "string",
            "enum": [
              "cloud_allowed",
              "local_only"
            ],
            "title": "Data Policy",
            "default": "cloud_allowed"
          },
          "preferred_profile": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 2,
                "pattern": "^[A-Za-z0-9_-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Profile"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          },
          "external_actions": {
            "type": "boolean",
            "const": false,
            "title": "External Actions",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "agent_id",
          "objective",
          "hypothesis_id"
        ],
        "title": "AgentRequest"
      },
      "AgentReview": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "choice": {
            "type": "string",
            "enum": [
              "accept",
              "reject"
            ],
            "title": "Choice"
          },
          "rationale": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 10,
            "title": "Rationale"
          },
          "expected_result_hash": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Expected Result Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "choice",
          "rationale",
          "expected_result_hash"
        ],
        "title": "AgentReview"
      },
...
      "Budget": {
        "properties": {
          "max_steps": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 12
          },
          "max_seconds": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Max Seconds",
            "default": 60
          },
          "max_cost_inr": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Max Cost Inr",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Budget"
      },
      "Command": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision"
        ],
        "title": "Command"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineCreate": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "template": {
            "type": "string",
            "enum": [
              "first_evidence",
              "pivot",
              "learning",
              "funding",
              "complete"
            ],
            "title": "Template"
          },
          "objective": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Objective"
          },
          "hypothesis_id": {
            "type": "string",
            "title": "Hypothesis Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "template",
          "objective",
          "hypothesis_id"
        ],
        "title": "PipelineCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}